+++ doc/changelog.rst
@@ -55,7 +55,7 @@ Release Notes
 - :bug:`admin,1579` There was a bug in the `pretalx init` command, and also too verbose output.
 - :bug:`orga,1577` The printable proposal cards showed broken characters for anything outside latin1.
 - :bug:`orga` Reordering questions while some of them were inactive could lead to 404 errors.
-- :bug:`orga:submission` pretalx wasn't able to filter out pending state changes from the organiser proposal list.
+- :bug:`orga:submission` pretalx wasn't able to filter pending state changes from the organiser proposal list.
 - :bug:`lang` The schedule editor was not operational with some languages, particularly with different language code versions (pt-BR vs pt_br).
 - :bug:`orga:schedule` The schedule editor would not show some specific time selectors when people expanded the timeline to see five-minute steps.
 - :release:`2023.1.0 <2023-08-30>`